+++ NEWS.md
@@ -150,6 +150,8 @@
   `win_current_group()` and `win_current_order()` are now exported. This
   should make it easier to provide customised SQL for window functions
   (#2051, #2126).
+  
+*  SQL translation for Microsoft SQL Server (@edgararuiz)
 
 ## Minor bug fixes and improvements